Move errors & export them
