+++ src/errors/custom-error.ts
