+++ test/errors/validation-error.test.ts
@@ -1,6 +1,6 @@
 import 'should';
 import * as sinon from 'sinon';
-import { ValidationError } from '../../../../src/middlewares/errors/validation-error';
+import { ValidationError } from '../../src/errors/validation-error';
 
 const sandbox = sinon.createSandbox();
